platform/tb-gateway/ota-transfer-service/requirements.txt
fastapi== 0.116.1
uvicorn==0.35.0
bleak==1.0.1
httpx==0.28.1
paho-mqtt== 2.1.0
